% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a place name or an address}
\usage{
geocode(string = "helsinki", hits = 1)
}
\arguments{
\item{string}{plane name or an address}

\item{hits}{Number of hits returned}
}
\value{
Data frame with latitude and longitude of the place
}
\description{
Geocode a place name or an address
}
\examples{
geocode(string = "Eduskuntatalo", hits = 1)
}
