% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode a geographical location}
\usage{
reverse_geocode(lon = "24.95126", lat = "60.15948", hits = 1)
}
\arguments{
\item{lon}{longitute}

\item{lat}{A latitude}

\item{hits}{Number of hits returned}
}
\value{
Data frame with information about the point
}
\description{
Reverse geocode a geographical location
}
\examples{
reverse_geocode(lon = "24.95126", lat = "60.15948", hits = 1)
}
